+++ src/main/java/com/example/auctionapp/repository/ItemRepository.java
@@ -22,7 +22,7 @@ public interface ItemRepository extends JpaRepository<Item, Long> {
     Item getByItemId(long itemId);
 
     @Query("SELECT i from Item i WHERE (i.category.supercategoryId = :supercategoryId or i.category.categoryId in :subcategories) and (i.startingPrice between :minPrice and :maxPrice)")
-    Page<Item> findItemsFiltered(@Param("supercategoryId")Long supercategoryId, @Param("subcategories")List<Long> subcategories, @Param("minPrice") Integer minPrice, @Param("maxPrice") Integer maxPrice, Pageable pageable);
+    Page<Item> findItemsFiltered(@Param("supercategoryId")Long supercategoryId, @Param("subcategories")List<Long> subcategories, @Param("minPrice") Long minPrice, @Param("maxPrice") Long maxPrice, Pageable pageable);
 
     List<Item> getByUserUserId(Long userId);
